requirements.txt
accelerate @ git+https://github.com/huggingface/accelerate@13ca7dccb6019e045a7cd2de9d0c832f6efd9aa0
aiofiles==23.2.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
dnspython==2.6.1
einops==0.8.0
email_validator==2.1.2
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
ffmpy==0.3.2
filelock==3.13.1
flash-attn==2.5.9.post1
fonttools==4.53.0
fsspec==2024.2.0
google-image-source-search==0.0.2
gradio==4.36.1
gradio_client==1.0.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
importlib_resources==6.4.0
ipykernel==6.29.4
ipython==8.24.0
ipywidgets==8.1.3
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
jsonnet-binary==0.17.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.1
nltk==3.8.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
orjson==3.10.5
packaging==24.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.1
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.7.4
pydantic_core==2.18.4
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.2
referencing==0.35.1
regex==2024.5.15
requests==2.31.0
rich==13.7.1
rpds-py==0.18.1
ruff==0.4.9
safetensors==0.4.3
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
starlette==0.37.2
sympy==1.12
textwrap3==0.9.2
tokenizers==0.19.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.3.1+cu121
torchaudio==2.3.1+cu121
torchvision==0.18.1+cu121
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
transformers @ git+https://github.com/huggingface/transformers@eed9ed679878ada2f6d2eefccdbda368cabc88b1
triton==2.3.1
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
wcwidth==0.2.13
websockets==11.0.3
widgetsnbextension==4.0.11
